+++ test/contract.js
@@ -668,11 +668,11 @@ var runTests = function(contractFactory) {
 
                 assert.deepEqual(result, {
                     contractAddress: address,
-                    cumulativeGasUsed: 10,
+                    cumulativeGasUsed: '10',
                     transactionIndex: 3,
                     blockNumber: 10,
                     blockHash: '0xbf1234',
-                    gasUsed: 0
+                    gasUsed: '0'
                 });
                 done();
             });
@@ -762,11 +762,11 @@ var runTests = function(contractFactory) {
             .on('receipt', function(result){
                 assert.deepEqual(result, {
                     contractAddress: address,
-                    cumulativeGasUsed: 10,
+                    cumulativeGasUsed: '10',
                     transactionIndex: 3,
                     blockNumber: 10,
                     blockHash: '0xbf1234',
-                    gasUsed: 0
+                    gasUsed: '0'
                 });
                 done();
             }).catch(console.log);
@@ -1899,12 +1899,12 @@ var runTests = function(contractFactory) {
                 // wont throw if it errors ?!
                 assert.deepEqual(receipt, {
                     contractAddress: null,
-                    cumulativeGasUsed: 10,
+                    cumulativeGasUsed: '10',
                     transactionIndex: 3,
                     transactionHash: '0x1234',
                     blockNumber: 10,
                     blockHash: '0x1234',
-                    gasUsed: 0,
+                    gasUsed: '0',
                     events: {
                         Unchanged: {
                             address: address,
@@ -2062,12 +2062,12 @@ var runTests = function(contractFactory) {
                     // wont throw if it errors ?! nope: causes a timeout
                     assert.deepEqual(receipt, {
                         contractAddress: null,
-                        cumulativeGasUsed: 10,
+                        cumulativeGasUsed: '10',
                         transactionIndex: 3,
                         transactionHash: '0x1234',
                         blockNumber: 10,
                         blockHash: '0x1234',
-                        gasUsed: 0,
+                        gasUsed: '0',
                         events: {
                             Changed: [
                                 {
@@ -2214,12 +2214,12 @@ var runTests = function(contractFactory) {
                     // wont throw if it errors ?!
                     assert.deepEqual(receipt, {
                         contractAddress: null,
-                        cumulativeGasUsed: 10,
+                        cumulativeGasUsed: '10',
                         transactionIndex: 3,
                         transactionHash: '0x1234',
                         blockNumber: 10,
                         blockHash: '0x43ffdd',
-                        gasUsed: 0,
+                        gasUsed: '0',
                         events: {
                             Unchanged: {
                                 address: address,
@@ -2329,12 +2329,12 @@ var runTests = function(contractFactory) {
                     if(count === 1) {
                         assert.deepEqual(receipt, {
                             contractAddress: null,
-                            cumulativeGasUsed: 10,
+                            cumulativeGasUsed: '10',
                             transactionIndex: 3,
                             transactionHash: '0x1234',
                             blockNumber: 10,
                             blockHash: '0x1234',
-                            gasUsed: 0,
+                            gasUsed: '0',
                             events: {}
                         });
 
@@ -2343,12 +2343,12 @@ var runTests = function(contractFactory) {
                     if(count === 2) {
                         assert.deepEqual(receipt, {
                             contractAddress: null,
-                            cumulativeGasUsed: 10,
+                            cumulativeGasUsed: '10',
                             transactionIndex: 3,
                             transactionHash: '0x1234',
                             blockNumber: 10,
                             blockHash: '0x1234',
-                            gasUsed: 0,
+                            gasUsed: '0',
                             events: {}
                         });