+++ test/formatters.inputTransactionFormatter.js
@@ -9,7 +9,7 @@ var tests = [{
         from: '0x11f4d0A3c12e86B4b5F39B213F7E19D048276DAe', // checksum address
         to: '0x11f4d0a3c12e86b4b5f39b213f7e19d048276dae',
         nonce: 1000,
-        gas: 1000,
+        gas: '1000',
         gasPrice: '1000'
     },
     result: {